+++ NEWS.rst
@@ -7,8 +7,8 @@ every change, see the Git log.
 Latest
 ------
 * Minor: Made ``peek`` const.
-* Minor: ``peek`` now takes a defaulted offset parameter for peeking further
-  ahead in the stream.
+* Minor: ``peek`` now takes an offset parameter for peeking further
+  ahead in the stream. This value is defaulted to 0 so the API hasn't changed.
 
 6.1.0
 -----