+++ NEWS.rst
@@ -6,6 +6,10 @@ every change, see the Git log.
 
 Latest
 ------
+* tbd
+
+6.2.0
+-----
 * Minor: Made ``peek`` const.
 * Minor: ``peek`` now takes an offset parameter for peeking further
   ahead in the stream. This value is defaulted to 0 so the API hasn't changed.